Begin Sys Inventory Rewrite
