+++ client/html/inventory/xp.js
